requirements.txt
aioredis==2.0.1
alembic==1.5.8
arq==0.20
asgiref==3.5.0
async-timeout==4.0.2
asyncpg==0.22.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cryptography==36.0.1
dependency-injector==4.38.0
dnspython==2.2.0
ecdsa==0.14.1
email-validator==1.1.3
fastapi==0.63.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
Mako==1.1.6
MarkupSafe==2.0.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.1
pyhumps==1.6.1
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
requests==2.27.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.11
starlette==0.13.6
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.1
setuptools